% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstabledist-package.R
\docType{package}
\name{dstabledist-package}
\alias{dstabledist}
\alias{dstabledist-package}
\title{dstabledist: The Discrete Stable Distribution Functions}
\description{
Probability generating function, formulae for the probabilities (discrete density) and random generation for discrete stable random variables.
}
\author{
\strong{Maintainer}: Krutto Annika \email{annikakrutto@gmail.com} (\href{https://orcid.org/0000-0001-7919-7683}{ORCID})

}
\keyword{internal}
